refactor(review): move the Node review digest to the tier-neutral lib
`src/app/review/digest.ts` called itself "the producer's hashing
implementation", but the daemon (`src/session/broker/state.ts`) imports it too,
which dragged a broker -> app-tier import edge across the codebase for three
lines of Node hashing. `src/lib/` already holds tier-neutral primitives
(`sourceText.ts`, `osPath.ts`, `terminalText.ts`), so the digest belongs there
and its header now says what it is rather than who owns it.
